Fix dtype_to_ctype for 64-bits
On some platforms (notably Windows), the size of a "long" may still
only be 32-bit. Because of this, the int64 and uint64 dtypes weren't
being mapped, despite running on a 64-bit host.

The "long long" ctype should always be 64-bit, so we can retain the
existing behaviour of using "long" when it is 64-bit, but support
platforms for which it isn't.